// src/models/diffusionRhItem.ts
export interface DiffusionRhItemI {
    id: number | null;
    lblOrganisation: string;
    situationActuelleAdministrationsCentrales: number | null;
    situationActuelleInteretsRegionaux: number | null;
    objectifsAtteindreYearOne: number | null;
    objectifsAtteindreYearTwo: number | null;
    objectifsAtteindreYearThree: number | null;
}